dom/base/test/test_bug382871.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=382871
-->
<head>
  <title>Test for Bug 382871</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=382871">Mozilla Bug 382871</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 382871 **/

function loadHandler(evt) {
  ok("randomProperty" in evt.target);
  ok("randomProperty" in evt.target.upload);
  SimpleTest.finish();
}

function runTest() {
  var xhr = new XMLHttpRequest();
  xhr.onload = loadHandler;
  xhr.randomProperty = true;
  xhr.upload.randomProperty = true;
  xhr.open("GET", "test_bug382871.html");
  xhr.send();
  xhr = null;
  SpecialPowers.gc();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);

</script>
</pre>
</body>
</html>
